% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addFragments}
\alias{addFragments}
\title{Add the restriction fragment information}
\usage{
addFragments(object, minSize = 20, filter = TRUE, save = FALSE)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{minSize}{Minimum size of a restriction fragment end to be valid.
Default is \code{20} bases.}

\item{filter}{Defines whether fragments that do not contain a cutting site
of the second restriction enzyme or are smaller than \code{minSize} should
be filtered out.}

\item{save}{Defines if the fragment information should be saved as txt and bed files
in the fragmentDir folder of the projectPath.}
}
\value{
Updated \code{FourC} object that contains the  information about
the restriction fragments in \code{rowData}.
}
\description{
\code{addFragments} adds the genomic position and information of the
restriction fragments as \code{GRanges} object to \code{rowData} of
the \code{FourC} object.
}
\examples{
exptData <- SimpleList(projectPath=tempdir(),
                       fragmentDir="re_fragments",
                       referenceGenomeFile=system.file("extdata/dm3_chr2L_1-6900.fa",
                                                       package="FourCSeq"),
                       reSequence1="GATC",
                       reSequence2="CATG",
                       primerFile=system.file("extdata/primer.fa",
                                              package="FourCSeq"),
                       bamFilePath=system.file("extdata/bam", package="FourCSeq"))

colData <- DataFrame(viewpoint = "testdata",
                     condition = factor(rep(c("WE_68h", "MESO_68h", "WE_34h"),
                                            each=2),
                                        levels = c("WE_68h", "MESO_68h", "WE_34h")),
                     replicate = rep(c(1, 2),
                                     3),
                     bamFile = c("CRM_ap_ApME680_WE_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_2_testdata.bam"),
                     sequencingPrimer="first")

fc <- FourC(colData, exptData)
fc

fc <- addFragments(fc)
fc
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{findViewpointFragments}}
}

