% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addPeaks}
\alias{addPeaks}
\title{Add peaks based on z-scores and adjusted p-values}
\usage{
addPeaks(object, zScoreThresh = 2, fdrThresh = 0.1)
}
\arguments{
\item{object}{\code{FourC} object for which z-scores have been calculated.}

\item{zScoreThresh}{Threshold on z-score that has to met by all replicates
of a condition.}

\item{fdrThresh}{Threshold on adjusted p-value that has to be met in at
least one replicated of a condition.}
}
\value{
\code{FourC} object with an additional assay 'peaks' of boolean
values. The parameters of the call to addPeaks are stored as DataFrame
\code{peakParameter} in the exptData slot.
}
\description{
Fragments are annotated to be a z-score peak (TRUE or FALSE) based
on z-scores and fdr values for the given fragment
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}

data(fcf, package="FourCSeq")

fcf <- addPeaks(fcf)
}

