% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addViewpointFrags}
\alias{addViewpointFrags}
\title{Add the information of the viewpoint fragments}
\usage{
addViewpointFrags(object, primerFragFile)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{primerFragFile}{character string defining the file that contains the
information about the primer fragments. Defaults to
\code{"primerFragments.rda"} which is the default output file
of \code{findViewpointFragments} in the provided \code{fragmentDir}.}
}
\value{
Updated \code{FourC} object with information about the viewpoint
fragments added to \code{colData}.
}
\description{
\code{addViewpointFrags} adds the genomic position and information of the
viewpoint primer fragments to \code{colData} of the \code{FourC} object.
}
\examples{
metadata <- list(projectPath=tempdir(),
                 fragmentDir="re_fragments",
                 referenceGenomeFile=system.file("extdata/dm3_chr2L_1-6900.fa",
                                                 package="FourCSeq"),
                 reSequence1="GATC",
                 reSequence2="CATG",
                 primerFile=system.file("extdata/primer.fa",
                                        package="FourCSeq"),
                 bamFilePath=system.file("extdata/bam", package="FourCSeq"))

colData <- DataFrame(viewpoint = "testdata",
                     condition = factor(rep(c("WE_68h", "MESO_68h", "WE_34h"),
                                            each=2),
                                        levels = c("WE_68h", "MESO_68h", "WE_34h")),
                     replicate = rep(c(1, 2),
                                     3),
                     bamFile = c("CRM_ap_ApME680_WE_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_2_testdata.bam"),
                     sequencingPrimer="first")

fc <- FourC(colData, metadata)
fc

fc <- addFragments(fc)

findViewpointFragments(fc)

fc <- addViewpointFrags(fc)
fc
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{findViewpointFragments}}
}

