% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countFragmentOverlapsSecondCutter}
\alias{countFragmentOverlapsSecondCutter}
\title{Count fragment overlaps when sequencing was performed from the second
cutting site}
\usage{
countFragmentOverlapsSecondCutter(object, extend = TRUE, minMapq = 0,
  shift = 0)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{extend}{Defines whether the read start should be extended by the length
of the second cutter recognition sequence to overlap the cutting sites which is
required for counting. If the cutting site has been trimmed with the primer
sequence this has to be set to \code{TRUE}. Default is \code{TRUE}.}

\item{minMapq}{Minimum mapping quality required for counting the read.
Default is \code{0}. If set to negative values the filtering step is skipped.}

\item{shift}{Maximum difference in starts or ends between read and fragment
positions. Default is \code{0}.}
}
\value{
Updated \code{FourC} object that contains two new \code{assays}
\code{countsLeftFragmentEnd} and \code{countsRightFragmentEnd} with the
count values at the respective fragment end.
}
\description{
\code{countFragmentOverlapsSecondCutter} counts the number of reads mapping to
each cutting site of the second cutter and then summarizes them over the
fragment ends of the first cutter stored in \code{rowRanges} of the
\code{FourC} object.
}
\examples{
## not run:
## countFragmentOverlapsSecondCutter(fc, extend=TRUE, minMapq=30)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{findViewpointFragments}},
\code{\link{countFragmentOverlaps}}
}

