% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distFitMonotone}
\alias{distFitMonotone}
\title{Fit the distance dependency}
\usage{
distFitMonotone(count, fragData, alpha = 20, penalty = 0.1,
  removeZeros = FALSE, ...)
}
\arguments{
\item{count}{Matrix of variance stabilized count data in the defined interval around
the viewpoint.}

\item{fragData}{Data frame with all the information on restriction
fragments and the interval around the viewpoint.}

\item{alpha}{Approximate number of fragments between two breaks of the B-spline
basis. \code{floor((number of fragments))/alpha)} is passed to \code{create.bspline.basis}
as \code{nbasis} argument (if this value is smaller than 4, 4 is the default value for \code{nbasis}).
Default is \code{20}.}

\item{penalty}{Penalty term passed to  \code{fdPar} as \code{lambda} argument.
Default is \code{0.1}.}

\item{removeZeros}{Defines if zero values should be removed from the fit.
Default is \code{FALSE}.}

\item{...}{Additional parameters.}
}
\value{
Returns a matrix with z-score values of all fragments around the given
viewpoint.
}
\description{
\code{distFitMonotone} takes the variance stabilized count values and
calculates a monotone fit for the distance dependency. The position
information about the viewpoint is stored in \code{fragData}.
The signal trend is fitted for the data left and right to the viewpoint
separately using the fda package.
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}

