% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distFitMonotoneSymmetric}
\alias{distFitMonotoneSymmetric}
\title{Fit the distance dependency}
\usage{
distFitMonotoneSymmetric(count, fragData, alpha = 20, penalty = 0.1,
  removeZeros = FALSE, ...)
}
\arguments{
\item{count}{Matrix of variance stabilized count data in the defined interval around
the viewpoint.}

\item{fragData}{Data frame with all the information on restriction
fragments and the interval around the viewpoint.}

\item{alpha}{Approximate number of fragments between two breaks of the B-spline
basis. \code{floor((max(number of fragments left or right))/alpha)} is passed to
\code{create.bspline.basis} as \code{nbasis} argument (if this value is smaller
than 4, 4 is the default value). Default is \code{20}.}

\item{penalty}{Penalty term passed to  \code{fdPar} as \code{lambda} argument.
Default is \code{0.1}.}

\item{removeZeros}{Defines if zero values should be removed from the fit.
Default is \code{FALSE}.}

\item{...}{Additional parameters.}
}
\value{
Returns a matrix with z-score values of all fragments around the given
viewpoint.
}
\description{
\code{distFitMonotoneSymmetric} takes the variance stabilized count values and
calculates a symmetric monotone fit for the distance dependency. The position
information about the viewpoint is stored in \code{fragData}.
The signal trend is fitted for the combined data left and right to the viewpoint
using the fda package.
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}

