\name{fc}
\alias{fc}
\docType{data}
\title{
FourC object with counts
}
\description{
This file contains the \code{FourCSeq} object \code{fc}. The reference fragments
have been added and the overlaps of the reference fragments and aligned reads
were counted. 
(See ?addFragments or the vignette of the 'FourCSeq' package for details).
}
\usage{data(fc)}
\format{
  Formal class 'FourC' [package "FourCSeq"]
}
\examples{
data(fc)
fc
}
\keyword{datasets}
