\name{fcf}
\alias{fcf}
\docType{data}
\title{
FourC object with z-scores
}
\description{
This file contains the \code{FourCSeq} object \code{fcf}. For this object
z-scores have been calculated. 
(See ?getZScores or the vignette of the 'FourCSeq' package for details).
}
\usage{data(fcf)}
\format{
  Formal class 'FourC' [package "FourCSeq"]
}
\examples{
data(fcf)
fcf
}
\keyword{datasets}