% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findViewpointFragments}
\alias{findViewpointFragments}
\title{Find the fragments to which the viewpoint primers map.}
\usage{
findViewpointFragments(object)
}
\arguments{
\item{object}{A \code{FourC} object.}
}
\description{
\code{findViewpointFragments} finds the position of the viewpoint primer in
the reference genome and on which restriction fragment they fall.
It saves the results in these files:
 - projectPath/fragmentDir/primerFragments.rda
 - projectPath/fragmentDir/primerFragments.txt
}
\examples{
exptData <- SimpleList(projectPath=tempdir(),
                       fragmentDir="re_fragments",
                       referenceGenomeFile=system.file("extdata/dm3_chr2L_1-6900.fa",
                                                       package="FourCSeq"),
                       reSequence1="GATC",
                       reSequence2="CATG",
                       primerFile=system.file("extdata/primer.fa",
                                              package="FourCSeq"),
                       bamFilePath=system.file("extdata/bam", package="FourCSeq"))

colData <- DataFrame(viewpoint = "testdata",
                     condition = factor(rep(c("WE_68h", "MESO_68h", "WE_34h"),
                                            each=2),
                                        levels = c("WE_68h", "MESO_68h", "WE_34h")),
                     replicate = rep(c(1, 2),
                                     3),
                     bamFile = c("CRM_ap_ApME680_WE_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_1_testdata.bam",
                                 "CRM_ap_ApME680_MESO_6-8h_2_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_1_testdata.bam",
                                 "CRM_ap_ApME680_WE_3-4h_2_testdata.bam"),
                     sequencingPrimer="first")

fc <- FourC(colData, exptData)
fc

fc <- addFragments(fc)

findViewpointFragments(fc)

fc <- addViewpointFrags(fc)
fc
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}
\code{\link{getReferenceSeq}}
}

