% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAllResults}
\alias{getAllResults}
\title{FourCSeq analysis results}
\usage{
getAllResults(object, ...)
}
\arguments{
\item{object}{\code{FourC} object for which z-scores and differences
have been calculated.}

\item{...}{Additional parameters that can be passed to \code{results}.}
}
\value{
DataFrame of results columns for all pairwise tests, with metadata
columns of coefficient and test information
}
\description{
Using the \code{DESEeq2} function \code{results},
\code{getAllResults} extracts results from a FourCSeq analysis
giving base means across samples,
log2 fold changes, standard errors, test statistics,
p-values and adjusted p-values for all pair-wise
conditions tested.
}
\examples{
data(fcf, package="FourCSeq")

fcf <- getDifferences(fcf, referenceCondition="WE_68h")

results <- getAllResults(fcf)
results
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link[DESeq2]{results}}
}

