% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDifferences}
\alias{getDifferences}
\title{Detect differences}
\usage{
getDifferences(object, referenceCondition = NULL, fitNormFactors = TRUE)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{referenceCondition}{Reference condition to be used for testing.}

\item{fitNormFactors}{Defines if the distance dependency should be used to
estimate normalizationFactors. Default is \code{TRUE}.}
}
\value{
Returns an updated \code{FourC} object that contains the results and
        information from the differential testing. Normalization factors
        are added to the assays slot.
}
\description{
\code{getDifferences} detects differences in the interaction frequencies between
different conditions. For each viewpoint all possible combinations of
conditions are tested using the \code{DESeq2} package.
}
\examples{
data(fcf, package="FourCSeq")

fcf <- getDifferences(fcf, referenceCondition="WE_68h")

results <- getAllResults(fcf)
results
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{getZScores}},
         \code{\link[DESeq2]{estimateDispersions}},
         \code{\link[DESeq2]{nbinomWaldTest}}
}

