% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDistAroundVp}
\alias{getDistAroundVp}
\title{getDistAroundVp}
\usage{
getDistAroundVp(vp, vpData, fragData)
}
\arguments{
\item{vp}{A single character string with the selected viewpoint.}

\item{vpData}{\code{DataFrame} containing the column data of the \code{FourC} object.}

\item{fragData}{\code{GRanges} object containing the row data (fragment data) of
the \code{FourC} object.}
}
\value{
Updated \code{GRanges} object that contains the distance from the viewpoint
        for the fragments on the viewpoint chromosome.
}
\description{
\code{getDistAroundVp} Add the distance information for a given viewpoint
to the fragment data.
}
\details{
Mid positions of the fragments are defined as start + (end - start) %/% 2
}
\examples{
data(fc, package="FourCSeq")

fragmentDataWithDistance <- getDistAroundVp("ap", colData(fc), rowData(fc))
}

