% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNormalizationFactors}
\alias{getNormalizationFactors}
\title{Get normalization factors for each fragment}
\usage{
getNormalizationFactors(object)
}
\arguments{
\item{object}{A \code{FourC} object.}
}
\value{
Returns an updated \code{FourC} object that contains normalization
factors.
}
\description{
\code{getNormalizationFactors} uses the distance dependency to calculate
normalization factors for each fragment.
}
\examples{
data(fcf, package="FourCSeq")

normalizationFactors <- getNormalizationFactors(fcf)
head(normalizationFactors)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{getDifferences}}
}

