% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getZScores}
\alias{getZScores}
\title{Calculate z-scores using the residuals of the general trend fit.}
\usage{
getZScores(object, removeZeros = TRUE, minCount = 40, minDist = NULL,
  fitFun = "distFitMonotoneSymmetric", sdFun = mad, ...)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{removeZeros}{Parameter to define if fragments with zero counts across
all replicates and conditions should be removed from the calculation.
Default is \code{TRUE}.}

\item{minCount}{The minimum median count value across replicates, a fragment
has to have to be used for the calculation of z-scores. Default is \code{40}}

\item{minDist}{Define a minimum distance to both sides of the viewpoint that is
discarded for analysis. If undefined (\code{NULL}), the borders of the
viewpoint peak are estimated as the minimum values next to the viewpoint before
the signal rises again.}

\item{fitFun}{Fit function to be used. Either \code{distFitMonotoneSymmetric} (default),
\code{distFitMonotone} or any self-defined function to fit the distance dependency.}

\item{sdFun}{Function to calculate the variation of the data. Default is \code{mad}.}

\item{...}{Additional Parameters passed to the distance fit function specified in
fitFun  (e.g. \code{distFitMonotone}, \code{distFitMonotoneSymmetric}).}
}
\value{
Returns a \code{FourC} object for the selected viewpoint
with z-score values for all fragments on the viewpoint chromosome that
passed the \code{minCount} threshold and that were not too close to the
viewpoint. All additional required data is saved in the object. Especially
the following information is added to the \code{FourC} object:
\enumerate{
     \item{metadata:} parameters passed to the getZScore function.
     \item{colData:} sdFun values calculated from the fit residuals
     \item{rowRanges:} the distance information to the viewpoint and information
                     calculated for the variance stabilizing data are added.
     \item{assays:} variance stabilized count values (trafo), fit values (fit),
                    z-score values (zScore), associated p-values (pValue) and
                    adjusted p-values (pAdjusted)
}
}
\description{
\code{getZScores} calculates the z-score for each fragment within a
distance from the viewpoint defined by \code{distAroundVP}. For the
calculation the count data is first transformed with a variance stabilizing
transformation from the \code{DESeq2} package. The decay trend of this transformed
data with distance from the viewpoint is fitted either with local regression
model using \code{locfit} or a monotone decay fit using the \code{fda} package.
z-scores are finally calculated from the residuals of the fit values.
}
\examples{
data(fc, package="FourCSeq")

fcf <- getZScores(fc)

fcf
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{countFragmentOverlaps}},
\code{\link{distFitMonotone}}, \code{\link{distFitMonotoneSymmetric}}
}

