% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalizeRPM}
\alias{normalizeRPM}
\title{Normalize count data to rpm}
\usage{
normalizeRPM(object, assay = "counts", normalized = "rpm",
  removeHighestValues = 0)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{assay}{Character vector selecting the assay of the \code{FourC} object
that should be normalized.}

\item{normalized}{Character vector specifying the assay name of the normalized
data.}

\item{removeHighestValues}{Defines the number of highest count values that should
       be removed before the library sizes are calculated in the denominator.
\code{normalized} containing the normalized data.}
}
\description{
\code{normalizeRPM} Normalizes the counts of each experiment to RPM.
}
\examples{
data(fc, package="FourCSeq")

fc <- combineFragEnds(fc)
fc

fc <- normalizeRPM(fc)
fc
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{combineFragEnds}}
}

