% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDifferences}
\alias{plotDifferences}
\title{Plot differences}
\usage{
plotDifferences(object, plotWindows = c(1e+05, 1e+06), textsize = 20,
  diffThresh = 0.01, controls = NULL, txdb = NULL, conditionAB = NULL)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{plotWindows}{Window sizes around the viewpoint for which plots are generated.}

\item{textsize}{Adjust text size.}

\item{diffThresh}{Threshold on adjusted p-values calculated in the differential test.}

\item{controls}{Auxiliary \code{GRanges} object that contains information about the
viewpoint and closest gene TSS.}

\item{txdb}{Auxiliary \code{TranscriptDb} object that contains gene models for the
investigated organism.}

\item{conditionAB}{Condition A and B, for which the comparison is plotted.}
}
\description{
\code{plotDifferences} generate plots to investigate the results
of \code{getDifferences}
}
\examples{
data(fcf, package="FourCSeq")

fcf <- getDifferences(fcf, referenceCondition="WE_68h")

plotDifferences(fcf)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{getDifferences}}
}

