% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotNormalizationFactors}
\alias{plotNormalizationFactors}
\title{Plot the estimated normalization factors.}
\usage{
plotNormalizationFactors(object, dist = TRUE, sizeFactors = TRUE)
}
\arguments{
\item{object}{A \code{FourC} object, after successfully calling \code{getZScores}.}

\item{dist}{If \code{TRUE} (default), the normalizationFactors are plotted as a
function for distance from the viewpoint. Otherwise they are plotted
as a function of ranks.}

\item{sizeFactors}{If \code{TRUE}, sizeFactors are calculated and plotted as
horizontal bars.}
}
\description{
\code{plotNormalizationFactors} generates plots of the fits used to calculate the z-scores.
}
\details{
Plots are generated to visualize the results of the fits used to calculate the z-scores.
}
\examples{
data(fcf, package="FourCSeq")

fcf <- getDifferences(fcf)

plotNormalizationFactors(fcf)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{getDifferences}},
        \code{\link[DESeq2]{estimateSizeFactors}}
}

