% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotScatter}
\alias{plotScatter}
\title{plotScatter}
\usage{
plotScatter(object, assay = "counts", ...)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{assay}{Character vector selecting the assay of the \code{FourC} object
for which the correlation between the replicates should be plotted.}

\item{...}{Additional arguments passed to \code{smoothScatter}.}
}
\description{
\code{plotScatter} Plot scatter plots for all experiments with replicates.
}
\examples{
data(fc, package="FourCSeq")

fc <- combineFragEnds(fc)
fc

plotScatter(fc)
}
\seealso{
\code{\link[graphics]{smoothScatter}}
}

