% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotZScores}
\alias{plotZScores}
\title{Plot z-score results.}
\usage{
plotZScores(object, cols = NULL, plotWindows = c(1e+05, 1e+06),
  controls = NULL, textsize = 20, txdb = NULL, plotSingle = FALSE)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{cols}{A character or numeric vector used to subset the columns of \code{object}.}

\item{plotWindows}{Window sizes around the viewpoint for which plots are generated.}

\item{textsize}{Plot parameter passed to the plotting function if}

\item{controls}{Auxiliary \code{GRanges} object that contains information about the
viewpoint and closest gene TSS.}

\item{txdb}{Auxiliary \code{TranscriptDb} object that contains gene models for the
investigated organism.}

\item{plotSingle}{If set to true each selected column will be plotted in a single plot.}
}
\description{
\code{plotZScores} generates plots to check the z-score calculation.
}
\details{
Plots are generated to visualize the results of the getZScores function
}
\examples{
data(fcf, package="FourCSeq")

plotZScores(fcf)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}, \code{\link{getZScores}},
\code{\link{distFitMonotone}}, \code{\link{distFitMonotoneSymmetric}}
}

