% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoothCounts}
\alias{smoothCounts}
\title{Smooth the counts of neighboring fragments}
\usage{
smoothCounts(object, assay = "counts", binWidth = 5)
}
\arguments{
\item{object}{A \code{FourC} object.}

\item{assay}{Assay name that will be smoothed.}

\item{binWidth}{Integer vector of odd numbers.}
}
\value{
Returns an updated \code{FourC} object with smoothed counts
for each binWidth as new assays.
}
\description{
Counts are smoothed using the number of fragments provided by binWidth.
binWidth has to be an odd number so that an equal number of fragments
to each side of the current fragment are used for smoothing.
}
\examples{
data(fc, package="FourCSeq")

fc <- smoothCounts(fc)
fc
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{FourC}}
}

