% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getNormalizationFactors}
\alias{getNormalizationFactors}
\title{Get normalization factors for each fragment}
\usage{
getNormalizationFactors(object)
}
\arguments{
\item{object}{A \code{FourC} object.}
}
\value{
Returns an updated \code{FourC} object that contains normalization
factors.
}
\description{
\code{getNormalizationFactors} uses the distance dependency to calculate 
normalization factors for each fragment.
}
\examples{
                       
data(fcf, package="FourCSeq")

normalizationFactors <- getNormalizationFactors(fcf)
head(normalizationFactors)

}
\seealso{
\code{\link{getDifferences}}
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
